docs(credits) Add JavaDocs
