+++ credits/src/main/java/com/ntt/microservice/credits/api/dto/request/CreditCardRequestDto.java
@@ -3,6 +3,9 @@ package com.ntt.microservice.credits.api.dto.request;
 import lombok.Getter;
 import lombok.Setter;
 
+/**
+ * Represents a request DTO for creating or updating a credit card.
+ */
 @Getter
 @Setter
 public class CreditCardRequestDto {
