+++ credits/src/main/java/com/ntt/microservice/credits/domain/repository/CreditCardRepository.java
@@ -10,9 +10,28 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface CreditCardRepository extends JpaRepository<CreditCard, String> {
+
+  /**
+   * Check if a credit card with the given customer ID exists.
+   *
+   * @param customerId The ID of the customer.
+   * @return true if a credit card with the given customer ID exists, false otherwise.
+   */
   boolean existsByCustomerId(String customerId);
 
+  /**
+   * Check if a credit card with the given card number exists.
+   *
+   * @param cardNumber The card number.
+   * @return true if a credit card with the given card number exists, false otherwise.
+   */
   boolean existsByCardNumber(String cardNumber);
 
+  /**
+   * Retrieve a credit card by customer ID.
+   *
+   * @param customerId The unique identifier of the customer.
+   * @return An Optional containing the CreditCard object, or empty if not found.
+   */
   Optional<CreditCard> findByCustomerId(String customerId);
 }
